osd: Deleting PG should discard pwlc
When a shard of a PG is deleted it clears the PG log and rolls
forward any updates that have not been completed, this cleans up
the object store removing any objects used for rollback. This
step updates pwlc.

Normally the next step is that the PG is destoryed, but there
is a (very rare) race hazard where a new interval and peering
cycle can start which will collect the updated pwlc from this
shard. This can disrupt the shards still being used by the PG.

Invalidating the pwlc after rolling forward updates fixes
this issue. As the PG is being deleted on this shard there is
no need to retain pwlc.

Fixes: https://tracker.ceph.com/issues/74048
Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>
(cherry picked from commit d78e0f22b6caee9efe99f2eebc5103a80f2c1e29)